requirements.txt
langgraph>=1.1.6
langchain-community>=0.4.1
langchain-ollama>=1.1.0
faiss-cpu>=1.13.2
openai>=1.30.0
google-generativeai>=0.8.0
pandas>=2.2.0
python-dotenv>=1.0.1
tqdm>=4.66.0
requests>=2.31.0
